Extract interface from SavedStateController
